% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOALDeR.R
\name{GOALDeR}
\alias{GOALDeR}
\title{The Generalized Outcome-Adaptive LASSO and Doubly Robust Estimate (GOALDeR)}
\usage{
GOALDeR(
  data,
  var.list,
  covar,
  Trt = "Trt",
  out = "Y",
  lambda_vec = c(-10, -5, -2, -1.5, -1.25, -1, -0.75, -0.5, -0.25, 0.25, 0.49),
  gamma_convergence = 2
)
}
\arguments{
\item{data}{Data that include outcome, treatment and all of covariates}

\item{var.list}{names of potential unknown confounders}

\item{covar}{names of known confounders}

\item{Trt}{Trt names of treatment}

\item{out}{out names of outcome}

\item{lambda_vec}{a vector of possible lambda values and defualt is c(-10,-5,-2,-1.5,-1.25,-1, -0.75, -0.5, -0.25, 0.25, 0.49)}

\item{gamma_convergence}{the value of gamma converagence and defualt is 2}
}
\value{
ATE is the linear dose-response function estimated using doubly robust estimator method.
}
\description{
The Generalized Outcome-Adaptive LASSO and Doubly Robust Estimate (GOALDeR)
}
