% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kennedy_etal.R
\name{kennedy_etal}
\alias{kennedy_etal}
\title{Doubly robust estimator by Kennedy et al.,2017}
\usage{
kennedy_etal(
  Y,
  A,
  X,
  Aseq,
  sl.lib = c("SL.glmnet", "SL.randomForest", "SL.xgboost", "SL.svm"),
  weights_est
)
}
\arguments{
\item{Y}{Outcome}

\item{A}{continuous treatment}

\item{X}{A matrix of covariates}

\item{Aseq}{continuous treatment}

\item{sl.lib}{methods combined by super learner and defualt is LASSO, XGBoost, Random Forest, and Support vector}

\item{weights_est}{balance weights}
}
\value{
lm_estimated is the estimated linear dose-response function
}
\description{
Doubly robust estimator by Kennedy et al.,2017
}
