% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_DR.R
\name{SL_DR}
\alias{SL_DR}
\title{Doubly robust estimator using super learner to fit the GPS model and the outcome model (SL-DR)}
\usage{
SL_DR(
  Y,
  A,
  X,
  treat_mod = c("Normal", "Gamma"),
  sl.lib = c("SL.glmnet", "SL.randomForest", "SL.xgboost", "SL.svm")
)
}
\arguments{
\item{Y}{Outcome}

\item{A}{continuous treatment}

\item{X}{A matrix of covariates}

\item{treat_mod}{marginal and conditional distributions of treatment}

\item{sl.lib}{methods combined by super learner and defualt is LASSO, XGBoost, Random Forest, and Support vector}
}
\value{
lm_estimated is the estimated linear dose-response function
}
\description{
Doubly robust estimator using super learner to fit the GPS model and the outcome model (SL-DR)
}
